+++ lib/goal_detail.dart
@@ -2,8 +2,7 @@ import 'package:flutter/material.dart';
 import 'constant.dart';
 import 'goal_delete.dart';
 import 'goal_update.dart';
-import 'task_edit_page.dart';
-
+import 'mission/task_edit_page.dart';
 
 class GoalDetailPage extends StatefulWidget {
   // 画面遷移元からのデータを受け取る変数
@@ -14,13 +13,11 @@ class GoalDetailPage extends StatefulWidget {
   _GoalDetailPageState createState() => _GoalDetailPageState();
 }
 
-
 class _GoalDetailPageState extends State<GoalDetailPage> {
   // 状態を管理する変数
   late int goalNumber;
   Goal? goalData;
   bool loading = false; // データをロード中かどうかを示すフラグ
-
 
   @override
   void initState() {
@@ -36,7 +33,7 @@ class _GoalDetailPageState extends State<GoalDetailPage> {
     });
 
     GoalResponse goalResponse =
-    await GoalResponse.fetchGoalResponse(goalNumber);
+        await GoalResponse.fetchGoalResponse(goalNumber);
     if (mounted) {
       setState(() {
         goalData = goalResponse.goal; // 新しいデータをリストに追加
@@ -44,7 +41,6 @@ class _GoalDetailPageState extends State<GoalDetailPage> {
       });
     }
   }
-
 
   @override
   Widget build(BuildContext context) {
@@ -62,167 +58,171 @@ class _GoalDetailPageState extends State<GoalDetailPage> {
       }
     }
 
-
     return Scaffold(
         appBar: AppBar(
           title: const Text('目標 詳細'),
         ),
-        body:SingleChildScrollView(
-          child:
-          Column(
-          children: <Widget>[
-            if (loading) Container(
-              alignment: Alignment.center,
-              padding: const EdgeInsets.all(16.0),
-              child: const CircularProgressIndicator(
-                strokeWidth: 3.0,
-              ),
-            ),
-            if (!loading)...[
+        body: SingleChildScrollView(
+          child: Column(
+            children: <Widget>[
+              if (loading)
+                Container(
+                  alignment: Alignment.center,
+                  padding: const EdgeInsets.all(16.0),
+                  child: const CircularProgressIndicator(
+                    strokeWidth: 3.0,
+                  ),
+                ),
+              if (!loading) ...[
                 Padding(
                   padding: const EdgeInsets.all(16.0),
-                  child:Column(
+                  child: Column(
                     children: <Widget>[
                       Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: <Widget>[
-                        Padding(
-                          padding: EdgeInsets.all(5),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text(
-                                '目標タイトル',
-                                style: const TextStyle(fontSize: 12.0, color: const Color(0xFFAE0103)),
-                              ),
-                              const SizedBox(height: 2.0),
-                              Text(
-                                '${goalData?.goalText}',
-                                style: const TextStyle(fontSize: 18.0),
-                              ),
-                            ],
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: <Widget>[
+                          Padding(
+                            padding: EdgeInsets.all(5),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: <Widget>[
+                                Text(
+                                  '目標タイトル',
+                                  style: const TextStyle(
+                                      fontSize: 12.0,
+                                      color: const Color(0xFFAE0103)),
+                                ),
+                                const SizedBox(height: 2.0),
+                                Text(
+                                  '${goalData?.goalText}',
+                                  style: const TextStyle(fontSize: 18.0),
+                                ),
+                              ],
+                            ),
                           ),
-                        ),
-
-                        Padding(
-                          padding: EdgeInsets.all(5),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text(
-                                '目的',
-                                style: const TextStyle(fontSize: 12.0, color: const Color(0xFFAE0103)),
-                              ),
-                              const SizedBox(height: 2.0),
-                              Text(
-                                '${goalData?.purpose}',
-                                style: const TextStyle(fontSize: 18.0),
-                              ),
-                            ],
+                          Padding(
+                            padding: EdgeInsets.all(5),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: <Widget>[
+                                Text(
+                                  '目的',
+                                  style: const TextStyle(
+                                      fontSize: 12.0,
+                                      color: const Color(0xFFAE0103)),
+                                ),
+                                const SizedBox(height: 2.0),
+                                Text(
+                                  '${goalData?.purpose}',
+                                  style: const TextStyle(fontSize: 18.0),
+                                ),
+                              ],
+                            ),
                           ),
-                        ),
-
-                        Padding(
-                          padding: EdgeInsets.all(5),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text(
-                                '目標達成時に得られるもの',
-                                style: const TextStyle(fontSize: 12.0, color: const Color(0xFFAE0103)),
-                              ),
-                              const SizedBox(height: 2.0),
-                              Text(
-                                '${goalData?.benefit}',
-                                style: const TextStyle(fontSize: 18.0),
-                              ),
-                            ],
+                          Padding(
+                            padding: EdgeInsets.all(5),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: <Widget>[
+                                Text(
+                                  '目標達成時に得られるもの',
+                                  style: const TextStyle(
+                                      fontSize: 12.0,
+                                      color: const Color(0xFFAE0103)),
+                                ),
+                                const SizedBox(height: 2.0),
+                                Text(
+                                  '${goalData?.benefit}',
+                                  style: const TextStyle(fontSize: 18.0),
+                                ),
+                              ],
+                            ),
                           ),
-                        ),
-
-
-                        Padding(
-                          padding: EdgeInsets.all(5),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text(
-                                '目標達成できなかった場合の損失',
-                                style: const TextStyle(fontSize: 12.0, color: const Color(0xFFAE0103)),
-                              ),
-                              const SizedBox(height: 2.0),
-                              Text(
-                                '${goalData?.loss}',
-                                style: const TextStyle(fontSize: 18.0),
-                              ),
-                            ],
+                          Padding(
+                            padding: EdgeInsets.all(5),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: <Widget>[
+                                Text(
+                                  '目標達成できなかった場合の損失',
+                                  style: const TextStyle(
+                                      fontSize: 12.0,
+                                      color: const Color(0xFFAE0103)),
+                                ),
+                                const SizedBox(height: 2.0),
+                                Text(
+                                  '${goalData?.loss}',
+                                  style: const TextStyle(fontSize: 18.0),
+                                ),
+                              ],
+                            ),
                           ),
-                        ),
-
-                        Padding(
-                          padding: EdgeInsets.all(5),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text(
-                                'メモ',
-                                style: const TextStyle(fontSize: 12.0, color: const Color(0xFFAE0103)),
-                              ),
-                              const SizedBox(height: 2.0),
-                              Text(
-                                '${goalData?.note}',
-                                style: const TextStyle(fontSize: 18.0),
-                              ),
-                            ],
+                          Padding(
+                            padding: EdgeInsets.all(5),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: <Widget>[
+                                Text(
+                                  'メモ',
+                                  style: const TextStyle(
+                                      fontSize: 12.0,
+                                      color: const Color(0xFFAE0103)),
+                                ),
+                                const SizedBox(height: 2.0),
+                                Text(
+                                  '${goalData?.note}',
+                                  style: const TextStyle(fontSize: 18.0),
+                                ),
+                              ],
+                            ),
                           ),
-                        ),
-
-                        Padding(
-                          padding: EdgeInsets.all(5),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text(
-                                '期限',
-                                style: const TextStyle(fontSize: 12.0, color: const Color(0xFFAE0103)),
-                              ),
-                              const SizedBox(height: 2.0),
-                              Text(
-                                '${goalData?.deadline.toString().substring(0, 10)} ${goalData?.deadline.toString().substring(11, 16)}',
-                                style: const TextStyle(fontSize: 18.0),
-                              ),
-                            ],
+                          Padding(
+                            padding: EdgeInsets.all(5),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: <Widget>[
+                                Text(
+                                  '期限',
+                                  style: const TextStyle(
+                                      fontSize: 12.0,
+                                      color: const Color(0xFFAE0103)),
+                                ),
+                                const SizedBox(height: 2.0),
+                                Text(
+                                  '${goalData?.deadline.toString().substring(0, 10)} ${goalData?.deadline.toString().substring(11, 16)}',
+                                  style: const TextStyle(fontSize: 18.0),
+                                ),
+                              ],
+                            ),
                           ),
-                        ),
-
-                        Padding(
-                          padding: EdgeInsets.all(5),
-                          child: Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text(
-                                '目標作成日時',
-                                style: const TextStyle(fontSize: 12.0, color: const Color(0xFFAE0103)),
-                              ),
-                              const SizedBox(height: 2.0),
-                              Text(
-                                '${goalData?.dateCreated.toString().substring(0, 10)} ${goalData?.dateCreated.toString().substring(11, 16)}',
-                                style: const TextStyle(fontSize: 18.0),
-                              ),
-                            ],
+                          Padding(
+                            padding: EdgeInsets.all(5),
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: <Widget>[
+                                Text(
+                                  '目標作成日時',
+                                  style: const TextStyle(
+                                      fontSize: 12.0,
+                                      color: const Color(0xFFAE0103)),
+                                ),
+                                const SizedBox(height: 2.0),
+                                Text(
+                                  '${goalData?.dateCreated.toString().substring(0, 10)} ${goalData?.dateCreated.toString().substring(11, 16)}',
+                                  style: const TextStyle(fontSize: 18.0),
+                                ),
+                              ],
+                            ),
                           ),
-                        ),
-                        SizedBox(height: 10.0),
-
-                      ],
-                    ),
+                          SizedBox(height: 10.0),
+                        ],
+                      ),
                       ElevatedButton(
                         onPressed: () {
                           Navigator.push(
                               context,
                               MaterialPageRoute(
-                                builder: (context) =>
-                                    GoalUpdatePage(),
+                                builder: (context) => GoalUpdatePage(),
                                 //builder: (context) => TaskEditPage(value: int.parse('352'))
                               ));
                         },
@@ -245,9 +245,9 @@ class _GoalDetailPageState extends State<GoalDetailPage> {
                           Navigator.push(
                               context,
                               MaterialPageRoute(
-                                builder: (context) =>
-                                    GoalDeletePage(value: int.parse(
-                                        '${goalData?.goalNumber}')),
+                                builder: (context) => GoalDeletePage(
+                                    value:
+                                        int.parse('${goalData?.goalNumber}')),
                                 //builder: (context) => TaskDeletePage(value: int.parse('352'))
                               ));
                         },
@@ -270,8 +270,6 @@ class _GoalDetailPageState extends State<GoalDetailPage> {
               ]
             ],
           ),
-        )
-    );
+        ));
   }
 }
-
