+++ lib/mission/mission_detail.dart
@@ -1,23 +1,22 @@
 import 'package:flutter/material.dart';
-import 'constant.dart';
+import 'mission_model.dart';
+import 'task_edit_page.dart';
+import 'mission_delete.dart';
 
-
-class MissionDeletePage extends StatefulWidget {
+class MissionDetailPage extends StatefulWidget {
   // 画面遷移元からのデータを受け取る変数
   final int value;
-  const MissionDeletePage({Key? key, required this.value}) : super(key: key);
+  const MissionDetailPage({Key? key, required this.value}) : super(key: key);
 
   @override
-  _MissionDeletePageState createState() => _MissionDeletePageState();
+  _MissionDetailPageState createState() => _MissionDetailPageState();
 }
 
-
-class _MissionDeletePageState extends State<MissionDeletePage> {
+class _MissionDetailPageState extends State<MissionDetailPage> {
   // 状態を管理する変数
   late int missionNumber;
   Mission? missionData;
   bool loading = false; // データをロード中かどうかを示すフラグ
-
 
   @override
   void initState() {
@@ -33,7 +32,7 @@ class _MissionDeletePageState extends State<MissionDeletePage> {
     });
 
     MissionResponse missionResponse =
-    await MissionResponse.fetchMissionResponse(missionNumber);
+        await MissionResponse.fetchMissionResponse(missionNumber);
     if (mounted) {
       setState(() {
         missionData = missionResponse.mission; // 新しいデータをリストに追加
@@ -41,7 +40,6 @@ class _MissionDeletePageState extends State<MissionDeletePage> {
       });
     }
   }
-
 
   @override
   Widget build(BuildContext context) {
@@ -59,47 +57,49 @@ class _MissionDeletePageState extends State<MissionDeletePage> {
       }
     }
 
-
     return Scaffold(
         appBar: AppBar(
-          title: const Text('Mission削除'),
+          title: const Text('Mission Detail'),
         ),
-        body:
-        Column(
+        body: Column(
           children: <Widget>[
-            if (loading) Container(
-              alignment: Alignment.center,
-              padding: const EdgeInsets.all(16.0),
-              child: const CircularProgressIndicator(
-                strokeWidth: 3.0,
+            if (loading)
+              Container(
+                alignment: Alignment.center,
+                padding: const EdgeInsets.all(16.0),
+                child: const CircularProgressIndicator(
+                  strokeWidth: 3.0,
+                ),
               ),
-            ),
-            if (!loading)...[
+            if (!loading) ...[
               Padding(
                 padding: const EdgeInsets.all(10.0),
-                child:
-                Column(
+                child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: <Widget>[
-                    Text(
-                      '以下のミッションを削除します。ミッション達成の記録も全て削除されますが、よろしいですか？',
-                      style: TextStyle(fontSize: 16.0),
-                    ),
-                    SizedBox(height: 12.0),
                     Text('・ミッションタイトル：${missionData?.missionText}'),
                     Text('・ご褒美：${missionData?.reward}'),
                     Text('・ペナルティ：${missionData?.penalty}'),
-                    Text('・開始日時：${missionData?.starTime.toString().substring(0, 10)} ${missionData?.starTime.toString().substring(11, 16)}'),
-                    Text('・終了日時：${missionData?.endTime.toString().substring(0, 10)} ${missionData?.endTime.toString().substring(11, 16)}'),
+                    Text(
+                        '・開始日時：${missionData?.starTime.toString().substring(0, 10)} ${missionData?.starTime.toString().substring(11, 16)}'),
+                    Text(
+                        '・終了日時：${missionData?.endTime.toString().substring(0, 10)} ${missionData?.endTime.toString().substring(11, 16)}'),
                     Text('・きっかけ：${missionData?.opportunity}'),
                     Text('・メモ：${missionData?.note}'),
-                    Text('・ミッション作成日時：${missionData?.dateCreated.toString().substring(0, 10)} ${missionData?.dateCreated.toString().substring(11, 16)}'),
+                    Text(
+                        '・ミッション作成日時：${missionData?.dateCreated.toString().substring(0, 10)} ${missionData?.dateCreated.toString().substring(11, 16)}'),
                   ],
-                ),),
+                ),
+              ),
               ElevatedButton(
                 onPressed: () {
-                  int count = 0;
-                  Navigator.popUntil(context, (_) => count++ >= 1);
+                  Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => TaskEditPage(
+                            value: int.parse('${missionData?.missionNumber}')),
+                        //builder: (context) => TaskEditPage(value: int.parse('352'))
+                      ));
                 },
                 style: ElevatedButton.styleFrom(
                   backgroundColor: const Color(0xFFAE0103),
@@ -108,7 +108,7 @@ class _MissionDeletePageState extends State<MissionDeletePage> {
                   ),
                 ),
                 child: const Text(
-                  'キャンセル',
+                  '編集',
                   style: TextStyle(
                     color: Colors.white,
                   ),
@@ -116,25 +116,29 @@ class _MissionDeletePageState extends State<MissionDeletePage> {
               ),
               ElevatedButton(
                 onPressed: () {
-                  int count = 0;
-                  Navigator.popUntil(context, (_) => count++ >= 2);
-                  // ここに削除の処理書く
+                  Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => MissionDeletePage(
+                            value: int.parse('${missionData?.missionNumber}')),
+                        //builder: (context) => TaskDeletePage(value: int.parse('352'))
+                      ));
                 },
-                child: const Text(
-                  '本当に削除する',
-                  style: TextStyle(
-                    color: Colors.white,
-                  ),
-                ),
                 style: ElevatedButton.styleFrom(
                   backgroundColor: const Color(0xFFAE0103),
                   shape: RoundedRectangleBorder(
                     borderRadius: BorderRadius.circular(10),
                   ),
                 ),
+                child: const Text(
+                  '削除',
+                  style: TextStyle(
+                    color: Colors.white,
+                  ),
+                ),
               ),
-            ]],
+            ]
+          ],
         ));
   }
 }
-
